# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/apps/noxum/properties": {
      "post": {
        "description": "Retrieves properties data from the specified URL",
        "operationId": "NoxumController_getProperties",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "URL to fetch properties data from",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "DUF user ID (required if client is dubaiunitfinder)",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "Properties data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "properties"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://example.com"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get properties data",
        "tags": [
          "noxum"
        ]
      }
    },
    "/apps/noxum/floorplans": {
      "get": {
        "description": "Retrieves floorplans data from the specified URL",
        "operationId": "NoxumController_getFloorplans",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "URL to fetch floorplans data from",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Floorplans data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "floorplans"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://example.com"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get floorplans data",
        "tags": [
          "noxum"
        ]
      }
    },
    "/apps/noxum/owners": {
      "get": {
        "description": "Retrieves owners data from the specified URL",
        "operationId": "NoxumController_getOwners",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "URL to fetch owners data from",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Owners data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "endpoint": {
                      "type": "string",
                      "example": "owners"
                    },
                    "url": {
                      "type": "string",
                      "example": "https://example.com"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get owners data",
        "tags": [
          "noxum"
        ]
      }
    },
...
    "/apps/noxum/combined/all": {
      "get": {
        "description": "Retrieves properties, floorplans, and owners data from the specified URL",
        "operationId": "NoxumController_getCombinedAll",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "URL to fetch combined data from",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Combined data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "success": {
                            "type": "boolean"
                          },
                          "endpoint": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get combined data (properties, floorplans, owners)",
        "tags": [
          "noxum"
        ]
      }
    },
    "/apps/noxum/combined/properties-owners": {
      "get": {
        "description": "Retrieves properties and owners data from the specified URL",
        "operationId": "NoxumController_getCombinedPropertiesOwners",
        "parameters": [
          {
            "name": "url",
            "required": true,
            "in": "query",
            "description": "URL to fetch combined data from",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Combined properties and owners data retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "success": {
                            "type": "boolean"
                          },
                          "endpoint": {
                            "type": "string"
                          },
                          "url": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Invalid scope"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get combined properties and owners data",
        "tags": [
          "noxum"
        ]
      }
    },
...
  },
  "info": {
    "title": "API Gateway",
    "description": "API Gateway with OAuth2.0 Client Credentials Flow",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:9000/v2",
      "description": "Production server"
    },
    {
      "url": "http://localhost:9000/v2",
      "description": "Development server"
    },
    {
      "url": "http://localhost:9000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}